syncs chart and app version on release
limits helm chart release to tags for a couple of reasons:
(1) if we change the chart, then we'll want another release and the
chart version should bump.
(2) if we change the image, then we'll want another release and the
chart should reflect the new image.
(3) it's easiser to reason about charts and image versions if they're
the same