adds istio sidecar ready/quit logic to crd commands
this may be the wrong place to put this logic. ideally, istio fixes
their shit and we no longer have to (a) wait for the sidecar to be ready
and (b) send an envoy quit signal after a job pod is done.

alas, this is not the case right now and wrapping this logic around the
actual command will require us to migrate off of a "distroless" base
docker image.

poop or vomit? pick one.