Adds support for running in Istio environments (#21)
* adds istio sidecar ready/quit logic to crd commands

this may be the wrong place to put this logic. ideally, istio fixes
their shit and we no longer have to (a) wait for the sidecar to be ready
and (b) send an envoy quit signal after a job pod is done.

alas, this is not the case right now and wrapping this logic around the
actual command will require us to migrate off of a "distroless" base
docker image.

poop or vomit? pick one.

* modifies helm chart

- adds support for '--istio-enabled' crd cmd flag
- adds psp use permissions to hooks
- adds custom psp for operator
- adds new values to control rendering

* adds network policy

exposed ports are health, metrics, and webhook. not sure we need to
block access to these but maybe i'm wrong

* disables codecov patch status check

not very important yet annoying